Working on UART Loading

+++ src/utils.h
@@ -1,6 +1,6 @@
 #ifndef _UTILS_H
 #define _UTILS_H
 
-void Busy_Wait(unsigned int wait);
+void Busy_Wait(unsigned long wait);
 
 #endif
